Copy of script with extras for kobra.io
This has functions that emulate Mudlet behavior, so people can test the script in a Lua terminal like the one at kobra.io. To use, just copy the contents of the file and paste it into the kobra.io script box. The original script is included in the middle, so don't paste it twice.